chore: フォルダ、ファイル名の変更（#20）
