+++ frontend/src/components/RegisterdFood/AllRegisteredFood.jsx
@@ -1,0 +1,5 @@
+export const AllRegisteredFood = () => {
+  return (
+    <h1>みんなの食品</h1>
+  );
+};
