--- frontend/src/pages/RegisterdFood.jsx
@@ -1,15 +1,0 @@
-import { FormProvider, useForm } from "react-hook-form";
-import { RegisterdFoodItem } from "../components/RegisterdFoodList/RegisterdFoodItem";
-
-export const RegisterdFood = () => {
-  const methods = useForm({
-    mode: "onBlur",
-    criteriaMode: "all"
-  });
-
-  return (
-    <FormProvider {...methods}>
-      <RegisterdFoodItem />
-    </FormProvider>
-  );
-};
